Migrate Mutation.addSlackIntegration
