+++ packages/services/api/src/modules/integrations/resolvers.ts
@@ -1,43 +1,11 @@
-import { z } from 'zod';
-import { HiveError } from '../../shared/errors';
 import { OrganizationManager } from '../organization/providers/organization-manager';
 import { IdTranslator } from '../shared/providers/id-translator';
 import type { IntegrationsModule } from './__generated__/types';
 import { GitHubIntegrationManager } from './providers/github-integration-manager';
 import { SlackIntegrationManager } from './providers/slack-integration-manager';
 
-/**
- * Current token size is 255 characters.
- * We allow some more for being future-proof :)
- *
- * https://api.slack.com/changelog/2016-08-23-token-lengthening
- */
-const SlackTokenModel = z.string().min(1).max(1000);
-
 export const resolvers: IntegrationsModule.Resolvers = {
   Mutation: {
-    async addSlackIntegration(_, { input }, { injector }) {
-      const AddSlackTokenIntegrationModel = z.object({
-        token: SlackTokenModel,
-      });
-
-      const result = AddSlackTokenIntegrationModel.safeParse(input);
-
-      if (!result.success) {
-        throw new HiveError(
-          result.error.formErrors.fieldErrors.token?.[0] ?? 'Please check your input.',
-        );
-      }
-
-      const organization = await injector.get(IdTranslator).translateOrganizationId(input);
-
-      await injector.get(SlackIntegrationManager).register({
-        organization,
-        token: input.token,
-      });
-
-      return true;
-    },
     async deleteSlackIntegration(_, { input }, { injector }) {
       const organizationId = await injector.get(IdTranslator).translateOrganizationId(input);
